server/chat_service/requirements.txt
Django<5.0,>=4.2
django-cors-headers
psycopg
    #to install REST_FRAMEWORK:
djangorestframework
markdown
django-filter 

    #install DAPHNE:
daphne

    #install channels:
channels
